+++ src/stelae/stele.rs
@@ -83,6 +83,7 @@ impl Stele {
         let dependencies = serde_json::from_str(&dependencies_str)?;
         Ok(Some(dependencies))
     }
+
     /// Get Stele's repositories.
     /// # Errors
     /// Will error if unable to find or parse repositories file at `targets/repositories.json`
@@ -96,6 +97,33 @@ impl Stele {
         let repositories_str = String::from_utf8(blob.content)?;
         let repositories: Repositories = serde_json::from_str(&repositories_str)?;
         self.repositories = Some(repositories.clone());
+        Ok(Some(repositories))
+    }
+
+    /// Check if Stele's private access file exists.
+    /// Returns true if `targets/private.json` exists, false otherwise.
+    #[must_use]
+    pub fn is_private_stelae(&self) -> bool {
+        self.auth_repo
+            .get_bytes_at_path("HEAD", "targets/private.json")
+            .is_ok()
+    }
+
+    /// Get Stele's repositories for specific commitish.
+    /// # Errors
+    /// Will error if unable to find or parse repositories file at `targets/repositories.json`
+    pub fn get_repositories_for_commitish(
+        &self,
+        committish: &str,
+    ) -> anyhow::Result<Option<Repositories>> {
+        let Ok(blob) = self
+            .auth_repo
+            .get_bytes_at_path(committish, "targets/repositories.json")
+        else {
+            return Ok(None);
+        };
+        let repositories_str = String::from_utf8(blob.content)?;
+        let repositories: Repositories = serde_json::from_str(&repositories_str)?;
         Ok(Some(repositories))
     }
 
@@ -140,12 +168,11 @@ impl Stele {
     /// Returns the first fallback repository found, or None if no fallback repository is found.
     #[must_use]
     pub fn get_fallback_repo(&self) -> Option<&Repository> {
-        self.repositories.as_ref().and_then(|repositories| {
-            repositories
-                .repositories
-                .values()
-                .find(|repository| repository.custom.is_fallback.unwrap_or(false))
-        })
+        let repositories = self.repositories.as_ref()?;
+        repositories
+            .repositories
+            .values()
+            .find(|repository| repository.custom.is_fallback.unwrap_or(false))
     }
 
     /// See if Stele is a root Stele.