+++ tests/api/mod.rs
@@ -1,4 +1,4 @@
+mod archive;
 mod archive_basic_test;
 mod archive_multihost_test;
 mod archive_multijursidiction_test;
-mod stelae;