+++ CHANGELOG.md
@@ -10,11 +10,19 @@ and this project adheres to a _modified_ form of _[Semantic Versioning][semver]_
 
 ### Added
 
+- Add `_archive` endpoint ([#81])
+- Restrict `_archive` endpoint access on private repositories ([#81])
+
 ### Changed
+
+- Bump rust-version to `1.88` ([86])
 
 ### Fixed
 
 ### Removed
+
+[#86]: https://github.com/openlawlibrary/stelae/pull/86
+[#81]: https://github.com/openlawlibrary/stelae/pull/81
 
 ## [0.5.1]